refactor: move navigation logic to separate component
