+++ App.js
@@ -1,25 +1,18 @@
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
+//REACT 
 import React from 'react';
-import {
-  Platform, StatusBar, StyleSheet, View
-} from 'react-native';
+import MainNavigation from './components/MainNavigation';
 
-// import useCachedResources from './modules/cached-resources/useCachedResources';
-import BottomTabNavigator from './navigation/BottomTabNavigator';
-import LinkingConfiguration from './navigation/LinkingConfiguration';
-import { initialize } from './services/parse/auth/index';
-import SignIn from './screens/SignIn';
-import SignUp from './screens/SignUp';
+//REDUX
+import { Provider } from "react-redux";
+import configureStore from './modules/state-management/configure-store';
 
+const store = configureStore();
 
-const Stack = createStackNavigator();
 
 export default class App extends React.Component {
-  constructor(props) {
-    super(props);
-    initialize();
-  }
+  // constructor(props) {
+  //   super(props);
+  // }
 
   render() {
     // const isLoadingComplete = useCachedResources();
@@ -28,23 +21,9 @@ export default class App extends React.Component {
     //   return null;
     // }
     return (
-      <View style={styles.container}>
-        {Platform.OS === 'ios' && <StatusBar barStyle="dark-content" />}
-        <NavigationContainer linking={LinkingConfiguration}>
-          <Stack.Navigator>
-            <Stack.Screen name="Sign In" component={SignIn} />
-            <Stack.Screen name="Sign Up" component={SignUp} />
-            <Stack.Screen name="Root" component={BottomTabNavigator} />
-          </Stack.Navigator>
-        </NavigationContainer>
-      </View>
+      <Provider store={store}>
+        <MainNavigation />
+      </Provider>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-  },
-});
