+++ components/MainNavigation/BottomTabNavigator./index.js
@@ -1,10 +1,10 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import * as React from 'react';
 
-import TabBarIcon from '../components/TabBarIcon';
-import HomeScreen from '../screens/HomeScreen';
-import DataCollection from '../screens/DataCollection';
-import DataAnalysis from '../screens/DataAnalysis';
+import TabBarIcon from '../../../components/TabBarIcon';
+import HomeScreen from '../../../screens/HomeScreen';
+import DataCollection from '../../../screens/DataCollection';
+import DataAnalysis from '../../../screens/DataAnalysis';
 
 const BottomTab = createBottomTabNavigator();
 const INITIAL_ROUTE_NAME = 'Home';
